requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
pymysql
pyjwt
cryptography

